+++ spec/factories/support/case.rb
@@ -8,5 +8,16 @@ FactoryBot.define do
     association :enquiry, factory: :support_enquiry
     association :category, factory: :support_category
     sub_category_string { "category subtitle" }
+
+    trait :with_documents do
+      transient do
+        document_count { 1 }
+      end
+
+      after(:create) do |kase, evaluator|
+        create_list(:support_document, evaluator.document_count, documentable: kase)
+        kase.reload
+      end
+    end
   end
 end